requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
pandas

redis==5.1.1
rq==1.16.2

httpx==0.27.2
openai==1.58.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
bcrypt==4.1.2

streamlit==1.39.0
